feat(develop): finish module task
